Add genre model, router, views.

+++ src/controllers/authorController.ts
@@ -1,7 +1,9 @@
 import express, { Response, NextFunction } from "express";
 import Author from "../models/author";
+import Book from "../models/book";
 import CustomRequest from '../middleware/customRequest';
 import log4js from "../middleware/logger";
+import CustomError from "../middleware/customError";
 
 const logger = log4js.getLogger("file");
 
@@ -14,7 +16,13 @@ class AuthorController {
     const user = req.user;
     try{
       const author = await Author.findById(req.params.id);
-      res.render("author/view", { author: author, user: user });
+      const booksIdList = author?.books || [];
+      const books = [];
+      for(let i = 0; i < booksIdList?.length; i++){
+        const book = await Book.findById(booksIdList[i]).populate("author").populate("genre").exec();
+        books.push(book);
+      }
+      res.render("author/view", { author: author, user: user, books: books });
     } catch(err){
       return next(err);
     } 
@@ -64,8 +72,12 @@ class AuthorController {
   public static deleteAuthor = async (req: CustomRequest, res: Response, next: NextFunction) => {
     const id = req.params.id;
     try{
-      const author = await Author.findOneAndDelete({ _id: id });
-      logger.info(`Author was successfully deleted: ${author}`);
+      const author = await Author.findOne({ _id: id });
+      if(author && author?.books.length > 0){
+        throw new CustomError(406, 'There are other books by this author in the library, you cannot delete the author until you delete all of his books')
+      }
+      const deletedAuthor = await Author.findOneAndDelete({ _id: id });
+      logger.info(`Author was successfully deleted: ${deletedAuthor}`);
       this.renderAuthor(req, res, next);
     } catch(err){
       return next(err);
